Separate out the DCAN HDF5 file creation.
