+++ .circleci/cifti_with_freesurfer_outputs.txt
@@ -24,10 +24,10 @@ xcp_d/sub-01/anat/sub-01_space-fsLR_den-32k_hemi-R_pial.surf.gii
 xcp_d/sub-01/anat/sub-01_space-fsLR_den-32k_hemi-R_smoothwm.surf.gii
 xcp_d/sub-01/func
 xcp_d/sub-01/func/sub-01_task-imagery_desc-filtered_motion.tsv
-xcp_d/sub-01/func/sub-01_task-imagery_desc-filtered_motion-DCAN.hdf5
+xcp_d/sub-01/func/sub-01_task-imagery_desc-dcan_qc.hdf5
 xcp_d/sub-01/func/sub-01_task-imagery_outliers.tsv
 xcp_d/sub-01/func/sub-01_task-imagery_run-01_desc-filtered_motion.tsv
-xcp_d/sub-01/func/sub-01_task-imagery_run-01_desc-filtered_motion-DCAN.hdf5
+xcp_d/sub-01/func/sub-01_task-imagery_run-01_desc-dcan_qc.hdf5
 xcp_d/sub-01/func/sub-01_task-imagery_run-01_design.tsv
 xcp_d/sub-01/func/sub-01_task-imagery_run-01_outliers.tsv
 xcp_d/sub-01/func/sub-01_task-imagery_run-01_space-fsLR_atlas-Glasser_den-91k_measure-pearsoncorrelation_conmat.pconn.nii
@@ -66,7 +66,7 @@ xcp_d/sub-01/func/sub-01_task-imagery_run-01_space-fsLR_den-91k_desc-denoisedSmo
 xcp_d/sub-01/func/sub-01_task-imagery_run-01_space-fsLR_den-91k_desc-denoisedSmoothed_bold.json
 xcp_d/sub-01/func/sub-01_task-imagery_run-01_space-fsLR_den-91k_qc.csv
 xcp_d/sub-01/func/sub-01_task-imagery_run-02_desc-filtered_motion.tsv
-xcp_d/sub-01/func/sub-01_task-imagery_run-02_desc-filtered_motion-DCAN.hdf5
+xcp_d/sub-01/func/sub-01_task-imagery_run-02_desc-dcan_qc.hdf5
 xcp_d/sub-01/func/sub-01_task-imagery_run-02_design.tsv
 xcp_d/sub-01/func/sub-01_task-imagery_run-02_outliers.tsv
 xcp_d/sub-01/func/sub-01_task-imagery_run-02_space-fsLR_atlas-Glasser_den-91k_measure-pearsoncorrelation_conmat.pconn.nii
