+++ xcp_d/utils/execsummary.py
@@ -1,6 +1,9 @@
 # emacs: -*- mode: python; py-indent-offset: 4; indent-tabs-mode: nil -*-
 # vi: set ft=python sts=4 ts=4 sw=4 et:
 """Functions for generating the executive summary."""
+from nipype import logging
+
+LOGGER = logging.getLogger("nipype.utils")
 
 
 def make_mosaic(png_files):
