+++ xcp_d/utils/qcmetrics.py
@@ -1,6 +1,11 @@
 """Quality control metrics."""
+import h5py
 import nibabel as nb
 import numpy as np
+import pandas as pd
+from nipype import logging
+
+LOGGER = logging.getLogger("nipype.utils")
 
 
 def compute_registration_qc(bold2t1w_mask, t1w_mask, bold2template_mask, template_mask):
@@ -196,3 +201,85 @@ def compute_dvars(datat):
     datax = np.hstack((firstcolumn, np.diff(datat)))
     datax_ss = np.sum(np.square(datax), axis=0) / datat.shape[0]
     return np.sqrt(datax_ss)
+
+
+def _make_dcan_qc_file(filtered_motion, TR):
+    """Make DCAN HDF5 file from single motion file."""
+    import os
+
+    from xcp_d.utils.execsummary import make_dcan_df
+
+    filtered_motion_filename = os.path.basename(filtered_motion)
+    dcan_df_file = f"{'.'.join(filtered_motion_filename.split('.')[:-1])}-DCAN.hdf5"
+    dcan_df_file = os.path.abspath(dcan_df_file)
+
+    make_dcan_df([filtered_motion], dcan_df_file, TR)
+    return dcan_df_file
+
+
+def make_dcan_df(fds_files, name, TR):
+    """Create an HDF5-format file containing a DCAN-format dataset.
+
+    Parameters
+    ----------
+    fds_files : list of str
+        List of files from which to extract information.
+    name : str
+        Name of the HDF5-format file to be created.
+    TR : float
+        Repetition time.
+
+    Notes
+    -----
+    FD_threshold: a number >= 0 that represents the FD threshold used to calculate
+    the metrics in this list.
+    frame_removal: a binary vector/array the same length as the number of frames
+    in the concatenated time series, indicates whether a frame is removed (1) or
+    not (0)
+    format_string (legacy): a string that denotes how the frames were excluded
+    -- uses a notation devised by Avi Snyder
+    total_frame_count: a whole number that represents the total number of frames
+    in the concatenated series
+    remaining_frame_count: a whole number that represents the number of remaining
+    frames in the concatenated series
+    remaining_seconds: a whole number that represents the amount of time remaining
+    after thresholding
+    remaining_frame_mean_FD: a number >= 0 that represents the mean FD of the
+    remaining frames
+    """
+    LOGGER.debug(f"Generating DCAN file: {name}")
+
+    # Load filtered framewise_displacement values from files and concatenate
+    filtered_motion_dfs = [pd.read_table(fds_file) for fds_file in fds_files]
+    filtered_motion_df = pd.concat(filtered_motion_dfs, axis=0)
+    fd = filtered_motion_df["framewise_displacement"].values
+
+    with h5py.File(name, "w") as dcan:
+        for thresh in np.linspace(0, 1, 101):
+            thresh = np.around(thresh, 2)
+
+            dcan.create_dataset(f"/dcan_motion/fd_{thresh}/skip", data=0, dtype="float")
+            dcan.create_dataset(
+                f"/dcan_motion/fd_{thresh}/binary_mask",
+                data=(fd > thresh).astype(int),
+                dtype="float",
+            )
+            dcan.create_dataset(f"/dcan_motion/fd_{thresh}/threshold", data=thresh, dtype="float")
+            dcan.create_dataset(
+                f"/dcan_motion/fd_{thresh}/total_frame_count", data=len(fd), dtype="float"
+            )
+            dcan.create_dataset(
+                f"/dcan_motion/fd_{thresh}/remaining_total_frame_count",
+                data=len(fd[fd <= thresh]),
+                dtype="float",
+            )
+            dcan.create_dataset(
+                f"/dcan_motion/fd_{thresh}/remaining_seconds",
+                data=len(fd[fd <= thresh]) * TR,
+                dtype="float",
+            )
+            dcan.create_dataset(
+                f"/dcan_motion/fd_{thresh}/remaining_frame_mean_FD",
+                data=(fd[fd <= thresh]).mean(),
+                dtype="float",
+            )
